Changed id type from int to Long

+++ src/main/java/com/acikgozKaan/LibraryAPI/dao/PublisherRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PublisherRepo extends JpaRepository<Publisher, Integer> {
+public interface PublisherRepo extends JpaRepository<Publisher, Long> {
 }
